Fix broken ipython doc links.
The ipython docs seem to no longer be available as scipy.org,
so switch the the equivalent link at ipython.org.

Things are further confused by the unification of the various
ipython backends under project jupyter. I moved the general
documentation link there, but the name change will be confusing.

Probably the whole section should be updated for the name change.